Updated CxImage library to the latest version (ours was a few years old).  This was needed to enable some of the more advanced processing functions I will need for some other investigations (particularly the DSP functions like histograms).

+++ agent/browser/ie/pagetest/PagetestReporting.cpp
@@ -2562,7 +2562,7 @@ void CPagetestReporting::CheckImageCompression()
 									img.SetJpegQuality(85);
 									target = size = bodyLen;
 									BYTE * mem = NULL;
-									long len = 0;
+									int len = 0;
 									if( img.Encode(mem, len, CXIMAGE_FORMAT_JPG) )
 									{
 										img.FreeMemory(mem);
